+++ lib/dollar.dart
@@ -1,3 +1,11 @@
 library dollar;
 
-export 'src/core.dart' show $handle, $effect, $cursor, $if, $Cursor;
+export 'src/core.dart' show $handle, $effect, $if, $ref, $Ref, $EffectHandler;
+export 'src/effects.dart'
+    show
+        $var,
+        $scan,
+        $Var,
+        $VarUpdateEffect,
+        $NeedCleanUpEffect,
+        $CleanedUpEffect;